Create abstract class to hold some common methods
