+++ backend/src/test/java/ChatsTest.java
@@ -3,10 +3,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Disabled;
 import com.bookbae.server.DatabasePoolService;
-import com.bookbae.server.SecretKeyService;
-import com.bookbae.server.service.SecretKeyServiceImpl;
-import com.bookbae.server.Login;
-import com.bookbae.server.CreateAccount;
 import com.bookbae.server.User;
 import com.bookbae.server.Like;
 import com.bookbae.server.Chats;
@@ -19,20 +15,13 @@ import com.bookbae.server.json.UserResponse;
 import com.bookbae.server.json.ChatCardResponse;
 import com.bookbae.server.json.ChatLineResponse;
 import java.util.UUID;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
 import java.util.ArrayList;
 import java.util.Random;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class ChatsTest {
+public class ChatsTest extends TestClass {
     private MockDatabaseService database;
-    private SecretKeyService keys;
-    private CreateAccount createAccountResource;
-    private Login loginResource;
-    private User userResource;
     private Like likeResource;
     private Chats chatsResource;
     private int numAccountRequests = 3; // arbitrarily set to 3
@@ -44,10 +33,6 @@ public class ChatsTest {
     @BeforeEach
     void init() {
         database = new MockDatabaseService("chatsTest");
-        createAccountResource = new CreateAccount(database);
-        keys = new SecretKeyServiceImpl();
-        loginResource = new Login(database, keys);
-        userResource = new User(database);
         likeResource = new Like(database);
         chatsResource = new Chats(database);
         database.init();
@@ -60,7 +45,7 @@ public class ChatsTest {
         userIds = new String[accountRequests.length];
 
         for (int i = 0; i < accountRequests.length; i++) {
-            userIds[i] = createMockUser(accountRequests[i], userRequests[i]);
+            userIds[i] = super.createMockUser(database, accountRequests[i], userRequests[i]);
         }
 
         //set 0th user and 1st user to like each other
@@ -141,61 +126,6 @@ public class ChatsTest {
         assertEquals(firstLine.getUserId(), userIds[0]);
         assertEquals(firstLine.getNthMessage(), 1);
 
-    }
-    // TODO: this is repeated in different forms throughout the tests, when we have time we should factor it out
-    // create an account with the given AccountRequest, put the User
-    private String createMockUser(AccountRequest accountRequest, UserRequest userRequest) {
-        var resp = createAccountResource.tryCreate(accountRequest);
-        assertEquals(200, resp.getStatus());
-        var acct = (LoginResponse) loginResource.tryLogin(accountRequest).getEntity();
-        var token = acct.getAuthToken();
-        Jws<Claims> jws;
-        try {
-            jws = Jwts.parserBuilder()
-                    .setSigningKey(keys.getKey())
-                    .build()
-                    .parseClaimsJws(token);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "";
-        }
-        String userId = jws.getBody().getSubject();
-        // assume ^ work
-
-        resp = userResource.putUser(new MockSecurityContext(userId), userRequest);
-        assertEquals(200, resp.getStatus());
-        return userId;
-    }
-
-    // creates a given number of accountRequests with random emails and passwords
-    private AccountRequest[] getExampleAccountRequests(int numRequests) {
-        AccountRequest[] acctRequests = new AccountRequest[numRequests];
-
-        for (int i = 0; i < acctRequests.length; i++) {
-            acctRequests[i] = new AccountRequest();
-            // UUID used here to create random email and password, not for user id
-            acctRequests[i].setEmail(UUID.randomUUID().toString() + "@example.com");
-            acctRequests[i].setPassword(UUID.randomUUID().toString());
-        }
-        return acctRequests;
-    }
-
-    // creates semi-random UserRequests based off of the given AccountRequests
-    private UserRequest[] getExampleUserRequests(AccountRequest[] acctRequests) {
-        UserRequest[] userRequests = new UserRequest[acctRequests.length];
-
-        for (int i = 0; i < userRequests.length; i++) {
-            userRequests[i] = new UserRequest();
-            userRequests[i].setEmail(acctRequests[i].getEmail());
-            userRequests[i].setName("Name " + i);
-            userRequests[i].setPreferredGender("M_F_NB");
-            userRequests[i].setGender("F");
-            userRequests[i].setFavGenre("Comedy");
-            userRequests[i].setBirthday("2000-01-01");
-            userRequests[i].setBio("User " + i + " bio");
-            userRequests[i].setZipcode((rand.nextInt(89999) + 10000) + ""); //10000 - 99999 zip codes
-        }
-        return userRequests;
     }
 
     private void doLike(String likerUserId, String likedUserId) {
