+++ backend/src/test/java/UserTest.java
@@ -4,8 +4,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Disabled;
 import com.bookbae.server.RestApplication;
 import com.bookbae.server.User;
-import com.bookbae.server.Login;
-import com.bookbae.server.CreateAccount;
 import com.bookbae.server.json.UserResponse;
 import com.bookbae.server.json.LoginResponse;
 import com.bookbae.server.json.UserRequest;
@@ -13,25 +11,18 @@ import jakarta.ws.rs.core.SecurityContext;
 import jakarta.ws.rs.core.Response;
 import com.bookbae.server.json.AccountRequest;
 import java.util.UUID;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import com.bookbae.server.service.SecretKeyServiceImpl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class UserTest {
+public class UserTest extends TestClass {
     private MockDatabaseService database;
     private User userResource;
-    private AccountRequest accountRequest;
-    private CreateAccount createAccountResource;
+    private AccountRequest accountRequest = super.getExampleAccountRequest();
 
     @BeforeEach
     void init() {
         database = new MockDatabaseService("userTest");
-        createAccountResource = new CreateAccount(database);
         userResource = new User(database);
-        accountRequest = getExampleAccountRequest();
         database.init();
     }
 
@@ -44,7 +35,7 @@ public class UserTest {
     @Test
     void getBasicUserTest() {
         // Create mock user
-        String userId = createMockUser();
+        String userId = super.createMockUser(database, false);
         assert !userId.equals("");
 
         // Get mock user and check if response email matches
@@ -57,7 +48,7 @@ public class UserTest {
     @Test
     void putUserTest() {
         // Create mock user
-        String userId = createMockUser();
+        String userId = super.createMockUser(database, false);
         assert !userId.equals("");
 
         // Put mock user
@@ -69,7 +60,7 @@ public class UserTest {
     @Test
     void putThenGetUserTest() {
         // Create mock user
-        String userId = createMockUser();
+        String userId = super.createMockUser(database, false);
         assert !userId.equals("");
 
         // Put mock user then get same user
@@ -93,54 +84,9 @@ public class UserTest {
 
     @Test
     void sqlFailureTest() {
-        createAccountResource.tryCreate(accountRequest);
         userResource = new User(new SQLFailService());
         var resp = userResource.getUser(new MockSecurityContext(UUID.randomUUID().toString()));
         assertEquals(500, resp.getStatus());
-    }
-
-    // create an account and get the user id of the created account
-    private String createMockUser() {
-        var keys = new SecretKeyServiceImpl();
-        createAccountResource.tryCreate(accountRequest);
-        var loginResource = new Login(database, keys);
-        var acct = (LoginResponse) loginResource.tryLogin(accountRequest).getEntity();
-        var token = acct.getAuthToken();
-        Jws<Claims> jws;
-        try {
-            jws = Jwts.parserBuilder()
-                    .setSigningKey(keys.getKey())
-                    .build()
-                    .parseClaimsJws(token);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "";
-        }
-        String subj = jws.getBody().getSubject();
-        // assume ^ work
-        return subj;
-    }
-
-    // creates AccountRequest with dummy data
-    private AccountRequest getExampleAccountRequest() {
-        var req = new AccountRequest();
-        req.setEmail("test@example.com");
-        req.setPassword("hunter2");
-        return req;
-    }
-
-    // creates UserRequest with dummy data
-    private UserRequest getExampleUserRequest() {
-        var req = new UserRequest();
-        req.setEmail("email@uw.edu");
-        req.setName("Name Name");
-        req.setPreferredGender("F_NB");
-        req.setGender("F");
-        req.setFavGenre("Horror");
-        req.setBirthday("2000-01-01");
-        req.setBio("bio");
-        req.setZipcode("80210");
-        return req;
     }
 
 }
